requirements.txt
Flask>=2.1.2
pytest>=7.1.2
marshmallow>=3.15.0
python-dotenv>=0.20.0
psycopg2-binary>=2.9.3
Flask-SQLAlchemy>=2.5.1
Flask-Migrate>=3.1.0
